api_yamdb/requirements.txt
requests==2.26.0
django==2.2.16
django-filter==2.3.0
djangorestframework==3.12.4
PyJWT==2.1.0
djangorestframework-simplejwt==4.7.2
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
gunicorn==20.0.4
psycopg2-binary==2.8.6